225
모든 warning 해결
aimparticle 만드는중
max count 제대로 설정 안되는것 해결 필요

+++ Engine_SOURCE/Entity.cpp
@@ -3,12 +3,13 @@
 
 namespace dru
 {
+
 	CEntity::CEntity()
-		: mID(reinterpret_cast<UINT32>(this))
+		: mID(reinterpret_cast<uintptr_t>(this))
 	{
 	}
 	CEntity::CEntity(const CEntity& other)
-		: mID(reinterpret_cast<UINT32>(this))
+		: mID(reinterpret_cast<uintptr_t>(this))
 		, mName(other.mName)
 	{
 	}
